ecs-django-microservices/app/requirements.txt
Django==4.2.16
django-admin-cli==0.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.1
python-dotenv==1.0.1
PyYAML==6.0.1